Refactor 1/2: moved detectors and verifiers to subfolders
